微信支付定时查询支付订单状态，超时修改本地订单状态 微信退款，查询退款信息

+++ pay-wechat/src/main/java/com/xiao/pay/paywechat/entity/PaymentInfo.java
@@ -54,7 +54,7 @@ public class PaymentInfo implements Serializable {
     /**
      * 支付金额(分)
      */
-    private Double payerTotal;
+    private Integer payerTotal;
 
     /**
      * 通知参数
